Generate only the correct number of imaginary registers.
The mechanism here is extremely cheesy, but it does work.
Each of the files defines one imaginary register as one plus the previous,
then includes the previous definition file. A linker script can provide just enough
definitions to mark out their zero page regions, then include the
appropriate file to fill out the rest. This is the simplest mechinism I could find
allowed this, but it's ripe for replacement the second a better idea comes
along.