Produce configuration files for Clang command lines.
This decouples the organization of the SDK from the contents of the
Clang driver. As an added bonus, the configuration files can provide
options as defaults that cannot easily be included in the Clang
toolchain description (-Os, -flto, --target=mos, etc.) This creates an
extremely straightforward out-of-the-box experience for using the
compiler as a whole. The defaults in the config files can easily be
overridden in the command line, and those overrides copy/pasted into new
configuration files, so this provides a nice smooth difficulty curve
when retargeting.