Refactor SDK for easier use and extensibility.
Uses ExternalProject to create a completely separate build directory
with cache configured for targeting MOS specifically. This is to avoid
the host cmake toolset interferring with the flags passed to mos clang
which can be particularly problematic on Windows.

Also generates clang symlinks for each platform added to the SDK. When
installed into the same prefix as clang, symlinks and config files are
installed alongside the clang binary. When added to PATH, each
platform's configuration is a single command invocation away. Any build
system can use this by prefixing clang with mos-<platform>-. For example:
`CC=mos-c64-clang`. Clang itself has the necessary logic to discover the
config file based on the argv0 prefix.

If a mos-clang is not detected by CMake, it will bootstrap llvm-mos from
the latest GitHub release by default.