Create an _init and _fini registration system.
This allows different snippets of initialization and finalization to be
linked in dynamically, based on what's actually used in the final
binary. The goal is to only pay size costs for functionality that is
actually used.

_start is redefined to be _init; each .init section is executed in
increasing numerical order upon startup. The .fini sections become a
_fini function that can be called from anywhere; as part of exit(), or
after main.

There's still a bit more compiler work necessary to elide the remainder
of the crt0, but we should be able to get it down pretty darn tight.