Fix case where atexit called within atexit handler.
"[Upon exit:] First, all functions registered by the atexit function are
called, in the reverse order of their registration, except that a
function is called after any previously registered functions that had
already been called at the time it was registered."

It's unspecified whether or not a call to atexit after exit() (i.e.,
inside an atexit() handler) succeeds, but we were summarily returning 1
even though the registered functions never get called.

Still, it's important that this call to __cxa_atexit succeed, since it
is necessary to clean up objects created in destructors. This is
accomplished by popping atexit handlers one at a time, which allows the
most recently popped handler to push handlers to be taken care of next.