[Major] Merge all nes-nrom targets together.
This add support for all (to the author's understanding) of the
additional ways of configuring the NROM mapper. This includes use of
CHR-RAM instead of CHR-ROM, NROM-128 vs NROM-256, and the use of PRG-RAM
as with Family Basic.

These options are now specified using the value of symbols, and the
linker script uses the value of these symbols to decide what to place
and where to place it.

This is a simple proof of concept for this approach; trying to do this
for MMC1 in a later change will be a more thorough trial.