[Major] Generalize MMC1 mapper support.
This fills out a generic linker script that provides indirect support
for all possible ways of laying out a board using the NES MMC1 mapper.
It's possible to place code and data in any bank, intended to be loaded
in any possible load address at runtime.

No support is provided for bank switching, ensuring the desired bank
mode is active at program start, etc. It's just a linker script; it
should be usable to build out almost arbitrarily complex setups, but we
leave this to the user, to avoid limiting their flexibility.