RAM shouldn't have a separate LMA region.
This was intended to be a way to track the bank of PRG-RAM values in a
mapper-independent way, but it just adds too much complexity to the
linker scripts, since RAM logically shouldn't have a LMA at all. This
caused the addresses to get mismatched, and its better to just not deal
with this.