Add C++ struct-of-arrays library and example
This adds two new types to the SDK under soa.h:
  - soa::Array<T, s> is an array of T of size s, but where the
    k'th byte of each T are stored contiguously.
  - soa::Ptr<T> is a pointer to an element of soa::Array.

The include-eable soa-struct.h specializes soa::Ptr<T> for a struct
type, including its members as distinct soa::Ptr<T> instances that
point to the members.

Using this library allows global arrays of structs to be accessed using
the absolute indexed addressing mode if it's always known at
compile-time what portion of the struct is being accessed. This is
generally true if there are no pointers into the array; in that common
case, using this library can produce a huge speedup.

See the included example for more details; the library is quite
expansive in practice, so the examples walk through what you can and
can't do with it.