[Breaking] Remove size-specific linker scripts from MMC1
MMC1 can now arrange PRG-ROM using only the value of the
`__prg_rom_size` symbol; it's no longer necessary to include a custom
linker script to set the size. Custom linker script inclusion is still
necessary to set the banking mode.

As part of this, the behavior of the 32-KiB banked mode is no longer
special-cased to initialize the bank to zero. This gives it the same
indeterminate startup behavior as the other banked modes; this
initialization must be done by hand.