Make PPU-only sets unreliable
`set` functions can be called in two places; inside VBLANK and outside.

If used outside, there's usually no way to observe from the CPU the
change in state, since the PPU is busy rendering. Accordingly, an
unreliable write should usually be fine, since it should only affect
rendering, not CPU behavior.

Inside VBLANK, there's not risk of being interrupted by a NMI unless the
VBLANK overflows, in which case the CPU may find itself outside VBLANK.
If not interrupted, a NMI-unreliable set is reliable, if interrupted,
then anything one would do with the PPU after the set isn't safe, since
an overrun has occurred. So, it seems like an unreliable semantics for
set is fine in this case too.