Rework banking logic entirely.
Man, not sure what I was thinking when I wrote this originally. Setting
the shadows after completing a write isn't safe; a NMI could occur in
between, and the shadow would be stale. A banked_call in the NMI handler
could then restore the stale state, and that would kill the write.

Each write also resets the shift register first, clobbering MMC1_CTRL.
Didn't notice that, sigh.

Instead, taking a page out of the other mappers, I've made the NMIs
summarily flush *all* shadow state. Then, a write to any register can
begin by flushing the shadow. If the write is interrupted, NMI will
reset the mapper and complete it, recording the interrupt as usual.
Then, the write can detect it was interruped, and just reset the shift
register and restore MMC1_CTRL. That works for any write, and the whole
model is just generally simpler with this. It looks like the fast path
for register writes is more efficient too; NMI gets a bit slower, having
to write PRG_BANK. All in all, seems worth it.