[Breaking] Rework MMC1 banking. (#204)
* [Breaking] Rework MMC1 banking.

This makes the interface for banking on MMC1 more consistent with the
other mappers.

- Removed direct access to shadow registers, replacing it with getter
  functions where required. These values may need to be computed on some
  mappers, so using functions allows the interfaces to look more similar
  across mappers. Inlining should completely fold this away.

- Added accessors for the high part of the CHR bank. On some MMC1
  boards, this controls PRG-RAM/ROM state, so it must be visibile to
  allow saving/restoring that state.

- Replaced the retry versions of chr bank setters with ones that
  reliably control just the high bits. The retry versions were always
  intended just to control the high bits; this makes this explicit and
  separate from true CHR bank control.

- The "set" functions that affect PPU/CHR state now take effect
  immediately, as well as being rewritten on NMI (previously they only
  did the latter). They are also made reliable; after return, a complete
  write has occurred. The "split" functions are left unreliable, since
  they should never conflict with NMI, but now internally disable IRQs.
  New "defer" functions were added to set the NMI state (the previous
  "set" behavior); these are intrinsically reliable.

- Remove weak from definitions; the library is fairly integrated, so it
  should be taken or left as a whole, pending a use for replacing
  specific functions.

- Make mirroring an enum type.

- Default NMI mirroring setting to vertical.

- Add routines to directly control CHR-ROM banking mode.

* Make PPU-only sets unreliable

`set` functions can be called in two places; inside VBLANK and outside.

If used outside, there's usually no way to observe from the CPU the
change in state, since the PPU is busy rendering. Accordingly, an
unreliable write should usually be fine, since it should only affect
rendering, not CPU behavior.

Inside VBLANK, there's not risk of being interrupted by a NMI unless the
VBLANK overflows, in which case the CPU may find itself outside VBLANK.
If not interrupted, a NMI-unreliable set is reliable, if interrupted,
then anything one would do with the PPU after the set isn't safe, since
an overrun has occurred. So, it seems like an unreliable semantics for
set is fine in this case too.

* Rename chr_rom_bank_mode setters to chr_bank_mode; they control RAM too

* Tail call in banked_call.

* WIP: Correct getter inconsistency

* Rework banking logic entirely.

Man, not sure what I was thinking when I wrote this originally. Setting
the shadows after completing a write isn't safe; a NMI could occur in
between, and the shadow would be stale. A banked_call in the NMI handler
could then restore the stale state, and that would kill the write.

Each write also resets the shift register first, clobbering MMC1_CTRL.
Didn't notice that, sigh.

Instead, taking a page out of the other mappers, I've made the NMIs
summarily flush *all* shadow state. Then, a write to any register can
begin by flushing the shadow. If the write is interrupted, NMI will
reset the mapper and complete it, recording the interrupt as usual.
Then, the write can detect it was interruped, and just reset the shift
register and restore MMC1_CTRL. That works for any write, and the whole
model is just generally simpler with this. It looks like the fast path
for register writes is more efficient too; NMI gets a bit slower, having
to write PRG_BANK. All in all, seems worth it.

* Remove dead function

* Fix typos

* Fix masks by making them symbolic

* Make split_mmc1_ctrl ignore PRG as advertised

* Remove bogus recursion limit