Use __builtin_bit_cast to make SoA constexpr if possible
reinterpret_cast is never allowed in constexpr, so this worked only but
for the grace of God. Using __builtin_bit_cast allows this to work in
more cases, but it still doesn't work if the data structure contains
pointers. There doesn't appear to be much we can do about this,
unfortunately.