Rewrite aligned_alloc
It appears more common for implementations to request a chunk large
enough to ensure a suitably-aligned sub-region and then to free the
unused portion. This is less likely to succeed, since it requires a
larger free space than strictly necessary, but it's simpler, and it
keeps malloc fast and small even if an aligned allocation is used, which
is more important. It will also be considerably faster if we get
constant time free list traversal.