Fix RP6502 varargs usage
All char varargs arguments are promoted to unsigned, so they must be
extracted as unsigned and casted back to char. The compiler should
optimize this away, but it's required to avoid undefined behavior.

It's also undefined behavior to pass the char count to va_start, but
fixing that wouldn't be transparent, and I think it should probably be
safe; we don't really use the argument to va_start.