[Breaking] Use shifted PETSCII by default
The spirit of the C standard is the implementation do whatever
translation is necessary to members of the execution character set to
have them stored or displayed properly on the target system. For
commodore targets, this really does mean that we should follow CC65 and
use the shifted character set.

Since our execution character set is always ASCII, we do need to swap
uppercase and lowercase characters at runtime. While the code to do this
isn't *that* horrible, its overhead is the main reason we haven't done
thus until now.

However, preliminary design work for stdio provided a way to do this by
default, while still allowing the user to turn it off. The default
__char_conv implementation for commodore does this translation and pulls
in a constructor that establishes shifted PETSCII at program start. If
the user provides their own __char_conv, the constructor will never be
pulled in, and they can use their own compile-time mechanisms to ensure
the printed strings are already in PETSCII as appropriate.