[Breaking] Switch to ISO mode for CX16
The spirit and letter of the C standard state that the effects of
printing printable characters in the execution character set (ASCII)
should produce the corresponding character on the display. Since the
Commander X16 has full ASCII support, we need to switch to ISO mode if a
character could ever be output to the screen. This is similar to the
shifted PETSCII switch now done for other Commodore family targets.