"[Breaking] Echo carraige return on commodore getchar"
This reverts commit 8c2e64cd8534885a392817548ca75df8b4daa3ec.

This is a tough one philosophically; doing this echo is extra code not
present in the KERNAL, and it *doesn't* traditionally belong to a libc.
However, the main high level programming language on these platforms is
BASIC, and BASIC does echo the return key, along with this being the
overwhelming behavior of modern line editors. It's also not much extra
code.

It seems like this is a case for pragmatism over idealism; we should
echo newlines on platforms where that's reasonably expected.