[Breaking] `freopen(NULL, ...)` now fails unconditionally
This is allowed by the C standard.  We were allocating a copy
of the filename from `fopen` just so `freopen` could be called
with NULL filename.  This is an unlikely scenario as `freopen`
is usually used for redirecting `stdout` etc.
In addition to wasted resources, there were multiple memory errors
managing the filename copy - see #356.

Fix #356