[atari] Buffer line on minimal getchar
The Atari OS expects that IOCB screen editor inputs are line-atomic; it
corrupts the cursor if writes happen in the meantime. It's generally
expected in C that one can interleave getchar and putchar, and that
putchars only take effect after a line is entered. Accordingly, this
change line-buffers the minimal getchar for Atari targets.

Fixes #378