Fix interruptable bank switch for NES-MMC3
If an interrupt occurs and the bank switch routine detects that
a retry is required, the orignal code has scratched the A register
at this stage and it causes a different bank to switch instead.
Using BIT in this manner shouldn't cause any incompatibilities due
to the possible locations for the RAM value (either in CIRAM or in
the mapper SRAM) cannot overlap with any registers on the console.