New global default stack size setting __set_heap_limit now sanity checks against stack; users can  now grab all memory for heap mos-sim flushes every output character -- no waiting for output
These changes were necessary to be able to get pi.c to calculate to 50000 digits on the Commodore 64.  Memory allocation had a tendency to walk all over the stack; this puts some sanity checks in place to prevent others from doing the same.  Default behavior is still 4 KB for heap.